requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-multipart==0.0.12

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.13.3

# HTTP Client
httpx==0.27.2

# Monitoring & Logging
sentry-sdk[fastapi]==2.17.0

# Resilience & Retry
tenacity==8.5.0

# Async Support
asyncio==3.4.3

# Utilities
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Scheduled Jobs (Phase 2.2)
apscheduler==3.10.4

# Data Processing (Phase 2+)
pandas==2.2.3
beautifulsoup4==4.12.3
aiosqlite==0.20.0
apscheduler==3.10.4
numpy==2.1.2

# Discord Integration (Phase 8)
discord.py==2.4.0

# Development & Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
black==24.10.0
ruff==0.7.2
mypy==1.13.0
